third_party/WebKit/LayoutTests/fast/dom/Range/getBoundingClientRect-expected.txt

Test 1
PASS rect.left.toFixed(3) is "8.000"
PASS rect.top.toFixed(3) is "8.000"
PASS rect.width.toFixed(3) is "400.000"
PASS rect.height.toFixed(3) is "400.000"
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height


Test 2
PASS rect.left.toFixed(3) is "8.000"
PASS rect.top.toFixed(3) is "452.000"
PASS rect.width.toFixed(3) is "400.000"
PASS rect.height.toFixed(3) is "376.000"
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height


Test 3
PASS rect.left.toFixed(3) is "8.000"
PASS rect.top.toFixed(3) is "1044.000"
PASS rect.width.toFixed(3) is "400.000"
PASS rect.height.toFixed(3) is "96.000"
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height


Test 4
PASS rect.left.toFixed(3) is "0.000"
PASS rect.top.toFixed(3) is "0.000"
PASS rect.width.toFixed(3) is "0.000"
PASS rect.height.toFixed(3) is "0.000"
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height


Test 5
PASS rect.left.toFixed(3) is "-14.574"
PASS rect.top.toFixed(3) is "1761.947"
PASS rect.width.toFixed(3) is "504.009"
PASS rect.height.toFixed(3) is "535.849"
PASS Math.abs(rect.left + rect.width - rect.right) < 0.001 is true
PASS Math.abs(rect.top + rect.height - rect.bottom) < 0.001 is true


Test 6
PASS rect.left.toFixed(3) is "0.000"
PASS rect.top.toFixed(3) is "0.000"
PASS rect.width.toFixed(3) is "0.000"
PASS rect.height.toFixed(3) is "0.000"
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height


PASS successfullyParsed is true

TEST COMPLETE
